+++ packages/acceptance-tests/pkg-tests-specs/sources/protocols/git.test.js
@@ -4,6 +4,7 @@ const {
 } = require(`pkg-tests-core`);
 const {parseSyml} = require(`@yarnpkg/parsers`);
 const {execUtils, semverUtils} = require(`@yarnpkg/core`);
+const {npath} = require(`@yarnpkg/fslib`);
 
 const TESTED_URLS = {
   // We've picked util-deprecate because it doesn't have any dependency, and
@@ -218,6 +219,30 @@ describe(`Protocols`, () => {
           } else {
             await expect(run(`install`)).rejects.toThrow(`Workspaces aren't supported by npm@${npmVersion}`);
           }
+        },
+      ),
+    );
+
+    test(
+      `it should not use Corepack to fetch Yarn Classic`,
+      makeTemporaryEnv(
+        {
+          dependencies: {
+            [`yarn-1-project`]: startPackageServer().then(url => `${url}/repositories/yarn-1-project.git`),
+          },
+        },
+        async ({path, run, source}) => {
+          // This checks that the `set version classic` part of `scriptUtils.prepareExternalProject` doesn't use Corepack.
+          // The rest of the install will fail though.
+          await expect(run(`install`, {
+            env: {
+              COREPACK_ROOT: npath.join(npath.fromPortablePath(path), `404`),
+              YARN_ENABLE_INLINE_BUILDS: `true`,
+            },
+          })).rejects.toMatchObject({
+            code: 1,
+            stdout: expect.stringContaining(`Saving the new release`),
+          });
         },
       ),
     );