+++ CHANGELOG.md
@@ -57,6 +57,7 @@ The following changes only affect people writing Yarn plugins:
 ### Bugfixes
 
 - `yarn dlx` will no longer report false-positive `UNUSED_PACKAGE_EXTENSION` warnings
+- When Corepack is enabled Yarn will now use the current CLI to prepare external Yarn classic projects, matching the behaviour of when Corepack is disabled.
 
 ## 3.2.2